% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df}
\alias{df}
\title{DYME data}
\format{
## `df`
A data frame with 1,199 rows and 11 columns of pregnant mice exposed to varying
 dyme doses in litters (clusters):
\describe{
  \item{no}{no malformation}
  \item{minor}{minor malformation}
  \item{yes}{full malformation}
  \item{id}{litter identification number}
  \item{dose}{administered dosage (numeric)}
  \item{dose_fct}{administered dosage (factor)}
  \item{weight}{weight of fetus}
  \item{litsz}{size of the litter}
  ...
}
}
\source{
<Price et al. (1987).>
}
\usage{
df
}
\description{
DYME data
}
\keyword{datasets}
