% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_aic_bic.R
\name{extract_aic_bic}
\alias{extract_aic_bic}
\title{Extract AIC, BIC and -2loglik of a glm/glmm model}
\usage{
extract_aic_bic(model)
}
\arguments{
\item{model}{model object
you wish to extract its AIC, BIC}
}
\value{
a data frame
}
\description{
Extract AIC, BIC and -2loglik of a glm/glmm model
}
\examples{
\dontrun{
extract_aic_bic()
}
}
