% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_overdispersion.R
\name{extract_overdispersion}
\alias{extract_overdispersion}
\title{Extract overdispersion parameter for a glm/glmm model}
\usage{
extract_overdispersion(model)
}
\arguments{
\item{model}{a glm/glmm model object}
}
\value{
a data frame
}
\description{
Extract overdispersion parameter for a glm/glmm model
}
\examples{
\dontrun{
extract_overdispersion()
}
}
