% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmsummary-package.R
\docType{package}
\name{glmsummary-package}
\alias{glmsummary}
\alias{glmsummary-package}
\title{glmsummary: Model Summary of GLM and GLMM family of Models}
\description{
Extraction of model summaries of GLM and GLMM family of models in R. Outputs model summary tables. Uses join functions in dplyr and piping to facilitate extraction.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bokola.github.io/glmsummary/}
}

}
\author{
\strong{Maintainer}: Basil Okola \email{okolabasilowiti@gmail.com}

}
\keyword{internal}
