% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_load_packages.R
\name{install_load_packages}
\alias{install_load_packages}
\title{Check for availability of required packages, install and load}
\usage{
install_load_packages(pkg)
}
\arguments{
\item{pkg}{a character vector of packages you wish to install}
}
\description{
Check for availability of required packages, install and load
}
\examples{
\dontrun{
install_load_packages()
}
}
