% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_estimates.R
\name{model_estimates}
\alias{model_estimates}
\title{Get model estimates, standard error and p-values}
\usage{
model_estimates(model)
}
\arguments{
\item{model}{an R object of the model}
}
\value{
a data frame of model coefficients, standard error and p-value
}
\description{
Get model estimates, standard error and p-values
}
\examples{
\dontrun{
model_estimates()
}
}
