% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_rank.R
\name{model_rank}
\alias{model_rank}
\title{Rank a data frame of models with respective AIC, BIC values}
\usage{
model_rank(x)
}
\arguments{
\item{x}{a data frame of model AIC, BIC, assumed to have many models to compare}
}
\value{
a ranked data frame of the models in terms of AIC, BIC
}
\description{
Rank a data frame of models with respective AIC, BIC values
}
\examples{
\dontrun{
model_rank()
}
}
