% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/working_corr.R
\name{working_corr}
\alias{working_corr}
\title{Get working correlation of a gee/glmm model}
\usage{
working_corr(model)
}
\arguments{
\item{model}{a gee/glmm model object}
}
\value{
a vector
}
\description{
Get working correlation of a gee/glmm model
}
\examples{
\dontrun{
working_corr()
}
}
